Fix return type of get_max_range() callback

+++ libharp/harp-ingestion-module.c
@@ -178,7 +178,7 @@ static int variable_definition_new(const char *name, harp_data_type data_type, i
                                    int (*read_all) (void *user_data, harp_array data),
                                    int (*read_range) (void *user_data, long index_offset, long index_length,
                                                       harp_array data),
-                                   int (*get_max_range) (void *user_data),
+                                   long (*get_max_range) (void *user_data),
                                    int (*read_sample) (void *user_data, long index, harp_array data),
                                    harp_variable_definition **new_variable_definition)
 {
@@ -822,7 +822,7 @@ harp_variable_definition *harp_ingestion_register_variable_full_read
 harp_variable_definition *harp_ingestion_register_variable_range_read
     (harp_product_definition *product_definition, const char *name, harp_data_type data_type, int num_dimensions,
      const harp_dimension_type *dimension_type, const long *dimension, const char *description, const char *unit,
-     int (*exclude) (void *user_data), int (*get_max_range) (void *user_data),
+     int (*exclude) (void *user_data), long (*get_max_range) (void *user_data),
      int (*read_range) (void *user_data, long index_offset, long index_length, harp_array data))
 {
     harp_variable_definition *variable_definition;
